requirements.txt
pandas 
numpy
seaborn
# matplotlib.pyplot
scikit-learn
flask
streamlit
python-dotenv
Pillow
streamlit

# google.generativeai
google-generativeai